Rename Currency class to EuropeanCurrency
After a bit of thought, it seemed clear that I was being ultra-specific
with the original class, and package, name, so much so and didn't need
to be. The grounds for that is that what I was ultimately checking for
was a European currency, not something specific to ING DiBa. As such,
the classes and package have been renamed appropriately.