Renamed class and package to be more intuitive
While I renamed the package in the previous commit, after reading
through the Zend Framework Filter introduction again,
https://docs.zendframework.com/zend-filter/intro/, I realised that I'd
not approached the package with the right mindset, and that the name of
the package and class were not correctly named. Given that, I've renamed
it again in the pursuit of having it be correct.